Increment version + remove config version
I also included error response for when I did not like returning null,
but I feel as though null is fine the way it is because we handle it,
while people adding placeholders only need to return it.

I removed config version because it was unused, and it is unlikely we
will add something that would break the current format.

No idea why gitignore was updated but oh well.