Fixes + Changes
RegEx clarification and fix for issues with legacy placeholders.
JavaScript service now supports value(...), which is the same as
get(...) except it returns a String instead of a Text, so you can check
for values (like value("%player_ping%")===10) more easily. Note: I have
not tested this change yet so I would hold off on using it until I push
the update to Ore.