Changes
Tidied up command exceptions.
Changed `Messages` to use an object instead of a string -> allows
calling `Messages.get().a.b.c.t(...)` or `Messages.get().a.b.c.value` to
get the parsed `Text` object or just a `String`.
Changed JavaScript to read files only on load/reload and not at all
times. This saves memory + allows files to be deleted while the server
is running, so that when you refresh the placeholder new scripts can be
present and old ones changed or removed.
3.10 should be ready to go with this update however I am unable to test
it right now because my copy of Sponge is not building in Eclipse.