Relational placeholders
Adds #21.
Service changes: added relational placeholder methods. Only does
anything if relational placeholders exist.
Expansion changes: added RelationalExpansion interface which specifies
two methods:
 onRelationalValueRequest(one, two, token) which returns an object.
Player one is parsed with priority. For example, if a player sends a
chat message, player one is sending it while player two is reading it.
 getSupportedRelationalTokens() which returns all supported relational
tokens.
By default, if you only implement RelationalExpansion, you do not have
to implement standard placeholders, but the option to support both
regular and relational exists.
Internal changes:
Config option to attempt to parse relational placeholders as recipient
based regular ones.
Messages updated to support relational info.
Parse command now considers the sender of the command, if it is a
player, as the recipient (player two) and parses for relational
placeholders.
/papi r now adds the internal placeholders, for example if you enable an
expansion and reload it should now be there. I have no way to officially
support doing this for any plugins implementing the API, so I have added
another method to the PlaceholderService:
registerReloadListener(Runnable, Optional<String>). This allows you to
register a runnable to be executed whenever the plugin is reloaded or,
if you provide a string in the optional, whenever that specific
placeholder is reloaded.

I have tested this and it SHOULD work, but if you find bugs, please open
an issue and I will try to resolve it.