Changes (API BREAKING!)
Changed implementation required method in Expansion from
onPlaceholderRequest to onValueRequest (to align with
RelationalPlaceholders and also to standardize and eliminate some
boilerplate code)
Internal fixes to support this change.

IF AN OBJECT THAT YOU RETURN DOES NOT SERIALIZE INTO TEXT PROPERLY,
PLEASE LET ME KNOW AND I WILL TRY TO WRITE IN SUPPORT FOR IT. (For now,
if I do not look for it, it will be serialized with .toString() and then
any &x color codes will be replaced.)

This next update will be quite the API breaking update lol. I think I
might make this PlaceholderAPI version 4.0 because of how much has
changed.