Refactoring
Moved all interfaces (including annotations) to main package, and all
other classes into impl subpackages.
Added ExpansionBuilder interface for better support for other
potentially implementing plugins. At some point I should test to make
sure I can replace this plugin with another one which does the same
thing. (I will eventually make Expansion an interface detailing what the
current class does, and rename the current ExpansionImpl, but until then
I will leave this as is.)
Better builder docs in one place.
Better builder creation through service.

In reality, it would make no sense for another plugin to implement the
service; it supports functions that are simply signatures of the code in
this plugin, and the code in this plugin is already here; it would be
like implementing SpongeCommon again. One big example is the annotations
system: sure, someone else could do it, but why go through the extra
effort to create something either (a): the exact same, using ASM
bytecode generation, which I should mention took A LONG TIME to figure
out, even with aids; or (b): slower than the current implementation, by
using reflection method calls, which can be 3 to 4 times slower than the
INVOKEVIRTUAL opcode (and for a plugin like this, where speed can be
important for parsing due to the fact that certain plugins using the
service may require placeholders to update 20 times per second, this
does make a difference. Of course, if the implementation of the
expansion is slow, it makes no difference, but for the default
placeholders and/or properly coded placeholders (that use caches and
high-speed calls), it is nice to be able to do this quickly.)

Because of this, I will likely not have true support for other
implementing plugins for some time, if ever, so I wouldn't expect much
change in the future moving towards that goal. My personal goal is to
offload as much as possible into the plugin, so that users of the
service can have minimal boilderplate code. Unfortunately for now, the
ExpansionBuilder is a bit of a step back, but I will try to provide
quick defaults in the future.

I think I type too much lol.