Proper null support
Turns out it wasn't just a minor oversight lol. From now on null source
and observer objects ARE accepted by the service, however not all
placeholders are guaranteed to parse. For placeholders, if you are
willing to accept null parameters, annotate them with @Nullable (see the
Defaults class for examples, though I do not have examples for nullable
sources or observers) to allow them to be null. If a parameter is NOT
annotated with nullable, a null check will be inserted for you so you
can safely assume any parameter not annotated with @Nullable is not
null. (For instance, a method with parameters Player, CommandSource and
String can use any of those parameters without checking if they are
null. If you annotate the CommandSource with nullable, you can check to
see if it is there or not.) It is important to note that PlaceholderAPI
will ignore Optional<String> tokens annotated with @Nullable and still
provide the token as an optional. There isn't much point to using that
annotation with it anyways, but that sanity check is there just in case.
This should fix the null issue.