Some useful changes
IExpansion interface added -> more of a technicality than anything else,
but it prevents exposure of Expansion class and instead allows an
interface to mask it. Good practice is all.
Requires works now. VersionRange stuff is good to go but additional
testing needs to be run first.
Issues still needing fixing: economy placeholder not allowing itself to
be enabled (attempting fix here but need to test on another system),
config resetting on restart in api 7.