Fixed issues with config resetting
What was happening was it was only resetting enabled values, but since
those were the only values in most config entries that was what was
changing. It was resetting that because I called a virtual check on the
data subnode instead of the node itself. Oops.
Fixed an issue where if a placeholder failed to load from defaults, they
all failed to load. Now if any placeholder fails to load it will just be
ignored.

Still need to make some changes to version dependency code because of
Sponge versioning but should be good to go soon.