Final fix + 4.3
Gonna push this update now. Should fix the current issues. From what I
can tell version checking does NOT work on in-IDE eclipse server
testing, but it seems to work just fine on my local test server so
hopefully it doesn't break anything. Not that it's used other than in
the statistics placeholder so whatever.